requirements/production.txt
-r base.txt

# Production monitoring
sentry-sdk[django]==1.40.0
newrelic==9.4.0

# Performance
django-cachalot==2.6.1
django-compressor==4.4

# Security
django-security==0.20.0
django-csp==3.7

# Health checks
django-health-check==3.17.0

# Backup
django-dbbackup==4.0.2
